+++ projects/ontimize-web-ngx/src/lib/interfaces/chart-on-demand.interface.ts
@@ -1,0 +1,3 @@
+export interface IChartOnDemandService {
+  openChartOnDemand(dataArray: any, service: any, entity: any);
+}